Move encrypter into test class
